moved boot_configuration concrete implementations to pixicorev6 package
